requirements.txt
# =============================================================================
# ARGUS - Python Dependencies
# =============================================================================

# Core async framework
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
aiosqlite>=0.19.0

# Exchange APIs
ccxt>=4.2.0
websockets>=12.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Telegram notifications
python-telegram-bot>=20.7

# Logging with colors
colorama>=0.4.6
colorlog>=6.8.0

# HTTP requests
requests>=2.31.0
httpx>=0.26.0

# Date/time handling
python-dateutil>=2.8.2
pytz>=2024.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Analysis (for Week 3)
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Database utilities
sqlite-utils>=3.36

# Type hints
typing-extensions>=4.9.0

# System tray (Windows background mode)
pystray>=0.19.5
pillow>=10.0.0

